SSPM: Merge single and multiple extended time slot
