+++ hardware/config/altde2-115-sspm.xml
@@ -12,8 +12,9 @@
   <Devs>
     <Dev DevType="SSPMDevice" entity="SSPM" iface="OcpCore">
       <params>
-      <param name="nCores" value="2"/>
-      <param name="extendedSlotSize" value="5"/>
+      <param name="nCores" value="2" />
+      <param name="extendedSlotSize" value="5" />
+      <param name="singleExtendedSlot" value="false" />
       </params>
     </Dev>
 	  <Dev DevType="Sram16" entity="SRamCtrl" iface="OcpBurst">
